refactor: decouple AsyncLocalStorage via DI
- Introduce ActionDispatchContext interface + DI token in common
- Dispatcher injects the context instance directly (toDynamicValue)
- Node app-module binds native AsyncLocalStorage from async_hooks
- Browser app-module binds AsyncLocalStorage from als-browser polyfill
- Keeps the common package runtime-neutral for browser bundling
