+++ packages/server/src/common/utils/promise-queue.spec.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022-2023 STMicroelectronics and others.
+ * Copyright (c) 2022-2026 STMicroelectronics and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,8 +14,10 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { delay } from '../test/mock-util';
-import { PromiseQueue } from './promise-queue';
+
 import { expect } from 'chai';
+// eslint-disable-next-line import-x/no-deprecated
+import { PromiseQueue } from './promise-queue';
 
 // Helper types and functions that are needed for test setup
 
@@ -74,11 +76,13 @@ function newTestPromise(resolveTime: number): TestPromise {
     return { state, promise };
 }
 
+// eslint-disable-next-line @typescript-eslint/no-deprecated, import-x/no-deprecated
 let queue = new PromiseQueue();
 
 // Test execution
 describe('test PromiseQueue', () => {
     beforeEach(() => {
+        // eslint-disable-next-line import-x/no-deprecated, @typescript-eslint/no-deprecated
         queue = new PromiseQueue();
     });
     it('enqueue - one element', async () => {
